+++ packages/lib/test/model/onChildAttachedTo.test.ts
@@ -44,6 +44,9 @@ beforeEach(() => {
         () => node,
         child => {
           const path = getParentToChildPath(node, child)!
+          if (!path) {
+            fail("path between " + node + " and " + child + " could not be found")
+          }
           log("attached", node, child, path)
           return () => {
             log("detached", node, child, path)