+++ packages/lib/test/parent/parent.test.ts
@@ -90,7 +90,7 @@ test("parent", () => {
     path: ["$", "arr", 0, "$"],
   })
 
-  expect(Array.from(getChildrenObjects(p).values())).toEqual([])
+  expect(Array.from(getChildrenObjects(p).values())).toEqual([p.arr, p.p2])
   expect(Array.from(getChildrenObjects(p.$).values())).toEqual([p.arr, p.p2])
   expect(Array.from(getChildrenObjects(p.arr).values())).toEqual(p.arr)
   expect(Array.from(getChildrenObjects(p.p2!).values())).toEqual([])